Move reader to internal package
