+++ cmd/opentelemetry/app/internal/reader/es/esspanreader/span_reader.go
