+++ cmd/opentelemetry/app/internal/reader/es/esspanreader/validate_query.go
