+++ cmd/opentelemetry/app/internal/reader/es/esspanreader/validate_query_test.go
